Add =net-fs/nfs-utils-1.3.1-r6
This provides a way to use the stable major version of net-fs/nfs-utils
with the current stable version of dev-libs/libevent.

Gentoo bug: https://bugs.gentoo.org/608152

nfs-utils-1.3.1-r6.ebuild and libevent.patch are from attachments to the
above bug. The remaining files are the same as in the main Portage tree.